+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Critical items to know are:
 versions here coincide with releases on pypi.
 
 ## [master](https://github.com/openbases/openbases-python/tree/master)
+ - adding badges module, and entry point ob-badge (0.0.3)
  - missing requests dependency (0.0.2)
  - adding icons endpoint `ob-icons` (0.0.1)
  - added base functions to get fields from paper markdown and tests (0.0.0)